fix(stage-15): make configuration and CI deterministic
Parse EMBEDDING_DIMENSIONS from .env while enforcing VECTOR(1024), run mypy in CI, align async integration fixtures with their event loop, and resolve existing typing/formatting drift. Verified with 108 unit/API tests and 8 PostgreSQL/NATS integration tests.

+++ src/tender_lens/crawler/base.py
@@ -112,9 +112,9 @@ class ResilientHttpClient:
                     if retry_at is not None:
                         if retry_at.tzinfo is None:
                             retry_at = retry_at.replace(tzinfo=UTC)
-                        return max(0.0, (retry_at - datetime.now(UTC)).total_seconds())
+                        return float(max(0.0, (retry_at - datetime.now(UTC)).total_seconds()))
         exponential = self._base_delay * (2 ** max(0, attempt - 1))
-        return exponential + self._jitter * self._random()
+        return float(exponential + self._jitter * self._random())
 
     async def _polite_delay(self) -> None:
         delay = self._base_delay + self._jitter * self._random()
